requirements.txt
flask
pymongo
feedparser
requests
python-dotenv
gunicorn
kafka-python
elasticsearch==8.15.1